requirements.txt
torch=2.5.0
torch-geometric
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
sam2
tensorboard
requests
tqdm
ogb
gitpython
opencv-python
scikit-learn
pillow
jupyter
torchvision
pycocotools
matplotlib
scikit-image
vtk
